Tester & Solution
This function-based Go program is run from the console, taking one user input, the number of prisoners. The program then initializes a slice containing the desired number of prisoners, assigning them each a random number, [1, number of prisoners]. This slice is printed to the console. Next, the solve() function is run, which implements the solution and checks if any prisoner correctly guesses their own number.